HevEventSourceTimeout: Fix timeout calculate.

+++ src/hev-event-source-timeout.c
@@ -59,7 +59,7 @@ hev_event_source_timeout_prepare (HevEventSource *source)
 	spec.it_interval.tv_sec = 0;
 	spec.it_interval.tv_nsec = 0;
 	clock_gettime (CLOCK_MONOTONIC, &spec.it_value);
-	time = spec.it_value.tv_sec * 1000000000 + spec.it_value.tv_nsec;
+	time = ((uint64_t) spec.it_value.tv_sec) * 1000000000 + spec.it_value.tv_nsec;
 	time += ((uint64_t) self->interval) * 1000000;
 	spec.it_value.tv_sec = time / 1000000000;
 	spec.it_value.tv_nsec = time % 1000000000;
